Works great, could add features and change things
SO, I stopped the first tutorial at about 2:30 in and decided to just do the program myself.
I'm not sure if this is what it's supposed to look like or do. 
I hope that I didn't teach myself any bad habits along the way. 
Got my rubocop offenses down below 40! 
1 file inspected, 39 offenses detected, 13 offenses autocorrectable
Here are some of my greatest hits: 
Assignment Branch Condition size for display_winning_line is too high.
Cyclomatic complexity for display_winning_line is too high.
Method has too many lines.
Line is too long. #==> I know, I know.
Use the return of the conditional for variable assignment and comparison.
      if active_player[1] == 'X'
Move break if open_squares.size == 0 out of the conditional (This one I was able to fix)
It's interesting to see what the errors are, and the style ones are pretty easy to address, but some of the functional ones are just over my head at this part of the curriculum, and I'm just glad that it all does what it's supposed to do and is consistent in its look and behavior. 
This was a fun project with lots of moving parts. Breaking everything down into smaller tasks made it doable.
Each task got its own mini pedac. Solving each problem was its own reward!

Thinking of adding functionality.
     should I track ties?
     add player list - rankings - extra players, all chosen computers 
         need player change breakout and leaderboard display
     add memory for playing all computer names. - major win if all beaten
     give each name a difficulty rating 
    (make 5 play styles? totally random to almost unbeatable)
     choice of random vs looking for block vs looking for win

Before adding any of that stuff, I'm going to go ahead and check out the tutorials
and see what I was SUPPOSED to be doing.