Works!
SO, I stopped the first tutorial at about 2:30 in and decided to just do the program myself.
I'm not sure if this is what it's supposed to look like or do. 
I apologize for submitting this out of order.
I hope that I didn't teach myself any bad habits along the way. 
Got my rubocop offenses down below 40! 
I still think that's pretty egregious.
1 file inspected, 36 offenses detected, 13 offenses autocorrectable
Here are some of my greatest hits: 
Assignment Branch Condition size for display_winning_line is too high.
Cyclomatic complexity for display_winning_line is too high.
Method has too many lines.
Line is too long. #==> I know, I know.
Use the return of the conditional for variable assignment and comparison.
      if active_player[1] == 'X'
It's interesting to see what the errors are, and the style ones are pretty easy to address,  
but some of the functional ones are just over my head at this part of the curriculum,   
and I'm just glad that it all does what it's supposed to do and is consistent in its look and behavior. 
I most likely have methods doing too many things at once.
And in some places they maybe shouldn't be methods at all.
A few mutate AND provide meaningful return values, which I know is not the way to do it.
This was a fun project with lots of moving parts. Breaking everything down into smaller tasks made it doable.
I really did my best to have a thorough initial pedac, and not hack and slash.
I definitely have more work to do there, though.
As the features grew, each task got its own mini pedac. Solving each problem was its own reward!
I left the method dissection in this time. It shouldn't really be there.
Thinking of adding functionality.
     should I use a YAML file to offload the dialog?
     should I track ties?
     add player list - rankings - extra players, all chosen computers 
         need player change breakout and leaderboard display
     add memory for playing all computer names. - major win if all beaten
     give each name a difficulty rating (the random sample technique is intentional because ties are the worst)
    (make 5 play styles? totally random to almost unbeatable)
     choice of random vs looking for block vs looking for win

Before adding any of that stuff, I'm going to go ahead and check out the tutorials
and see what I was SUPPOSED to be doing.
I'm feeling pretty sophomoric and want to get past that.
Thank you to anyone who checks this out.